New possible Fulmen logo
The logo is one of Zeus' symbols. Being Zeus the greek god of thunders, it fits well with the app theme. This is just a mix of a 4 and a 2 made on MS Paint, because for the moment i didn't manage to find a high quality image of the symbol on the web.